Showing nonheir scene

+++ PhongMaterial.cpp
@@ -1,7 +1,7 @@
 #include "PhongMaterial.hpp"
 
 PhongMaterial::PhongMaterial(
-    const glm::vec3& kd, const glm::vec3& ks, double shininess )
+    const glm::dvec3& kd, const glm::dvec3& ks, double shininess )
     : m_kd(kd)
     , m_ks(ks)
     , m_shininess(shininess)
